:hammer: simplify otlp handler
